initial commit
A NodeJS gRPC client communicating over TLS to a Golang gRPC server. The
server is configured with a very short 1 millisecond connection timeout.
Even though the client is never fast enough to finish the TLS handshake
within the timeout, gprc-js will (after a few tries) have a Http2Session
object which is reporting a live connection.